+++ packages/xml-views-completion/test/providers/attributeValue/enum-spec.ts
@@ -1,15 +1,22 @@
 import { expect } from "chai";
 import { forEach, map } from "lodash";
 import { UI5SemanticModel } from "@ui5-language-assistant/semantic-model-types";
-import { generateModel } from "@ui5-language-assistant/test-utils";
+import {
+  generateModel,
+  GEN_MODEL_TIMEOUT
+} from "@ui5-language-assistant/test-utils";
 import { testSuggestionsScenario } from "../../utils";
 import { enumSuggestions } from "../../../src/providers/attributeValue/enum";
 import { UI5XMLViewCompletion } from "../../../api";
 import { XMLAttribute, XMLElement } from "@xml-tools/ast";
 
-const ui5SemanticModel: UI5SemanticModel = generateModel("1.74.0");
-
 describe("The ui5-language-assistant xml-views-completion", () => {
+  let ui5SemanticModel: UI5SemanticModel;
+  before(async function() {
+    this.timeout(GEN_MODEL_TIMEOUT);
+    ui5SemanticModel = await generateModel({ version: "1.74.0" });
+  });
+
   context("enum values", () => {
     context("applicable scenarios", () => {
       it("will suggest enum values with no prefix provided", () => {