+++ packages/vscode-ui5-language-assistant/CHANGELOG.md
@@ -3,6 +3,20 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.2](https://github.com/sap/ui5-language-assistant/tree/master/packages/vscode-ui5-language-assistant/compare/vscode-ui5-language-assistant@1.1.1...vscode-ui5-language-assistant@1.1.2) (2020-04-07)
+
+**Note:** Version bump only for package vscode-ui5-language-assistant
+
+## [1.1.1](https://github.com/sap/ui5-language-assistant/tree/master/packages/vscode-ui5-language-assistant/compare/vscode-ui5-language-assistant@1.1.0...vscode-ui5-language-assistant@1.1.1) (2020-04-07)
+
+### Bug Fixes
+
+- **vscode-ui5-language-assistant:** full schema v1.19 ([#49](https://github.com/sap/ui5-language-assistant/tree/master/packages/vscode-ui5-language-assistant/issues/49)) ([b5592c3](https://github.com/sap/ui5-language-assistant/tree/master/packages/vscode-ui5-language-assistant/commit/b5592c307a86d72408463868b218ef60989c2ff0))
+
+### Performance Improvements
+
+- **language-server:** cache downloaded resources ([#50](https://github.com/sap/ui5-language-assistant/tree/master/packages/vscode-ui5-language-assistant/issues/50)) ([de8d7d5](https://github.com/sap/ui5-language-assistant/tree/master/packages/vscode-ui5-language-assistant/commit/de8d7d5e38c76666cc2590a885127b202096f289))
+
 # [1.1.0](https://github.com/sap/ui5-language-assistant/tree/master/packages/vscode-ui5-language-assistant/compare/vscode-ui5-language-assistant@1.0.0...vscode-ui5-language-assistant@1.1.0) (2020-04-01)
 
 ### Features